git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@4068 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/jd/plugins/host/QshareCom.java
@@ -220,7 +220,7 @@ public class QshareCom extends PluginForHost {
         String[] dat = br.getRegex("<SPAN STYLE=\"font-size:13px;vertical-align:middle\">(.*?) \\((.*?)\\).*?</SPAN>").getRow(0);
         if (dat.length != 2) throw new PluginException(LinkStatus.ERROR_FILE_NOT_FOUND);
         downloadLink.setName(dat[0].trim());
-        downloadLink.setDownloadSize((int) Regex.getSize(dat[1].trim()));
+        downloadLink.setDownloadSize(Regex.getSize(dat[1].trim()));
         return true;
     }
